+++ frontend/src/features/dashboard/components/Sidebar.tsx
